working on audio menu & int instead of float for night number

+++ Assets/Scripts/DebugFunctions.cs
@@ -2,7 +2,7 @@
 
 public class DebugFunctions : MonoBehaviour
 {
-    private float nightNumber;
+    private int nightNumber;
 
     SaveGameState saveGameState;
     SaveManager saveManager;
